Make the `css` function use args, not take an array.
Summary: Instead of taking an array, we can just use the list of arguments that
was taken. The `...rest` operator makes this really easy.

Also, clean up the `css()` tests to be a lot nicer. And add a `.gitattributes`
file so that the dist/ file is treated as binary.

Test plan:
 - `npm run test`

Auditors: jlfwong