Add in support for @font-face rules so that they are globally available.
Summary:
This adds in support for the global `@font-face` rule, inserting it explicitly on StyleSheet creation. This is a weird case because it's not directly associated with an element on the page. I figure having it exist on StyleSheet creation is OK as if the module isn't loaded then the font-face rule won't be inserted into the page.

I added in support for this so that a FontAwesome implementation could be built.

Test Plan:
Use it and check to see that custom fonts work. They do! Example:
https://www.dropbox.com/s/v3f24v31yk13vy4/Screenshot%202015-10-26%2020.58.50.png?dl=0

Reviewers: emily, jlfwong

Reviewed By: jlfwong

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D22756