Make the suffix of class names deterministic based on their values.
Summary:
Use a hash of the styles for each class name to determine the suffix
added to the name. This means that if multiple styles are created in different
places with the same values, they will use the same class. This is mostly
useful for ensuring that the suffixes are generated the same on both servers
and clients.

I'm still not 100% sure that `JSON.stringify` necessarily produces the same
results on all clients. It looks like ES2015 compliant browser engines have
defined property order that we can count on, but I'm not sure if there are any
of those yet? But Ben Alpert has made a convincing case that Facebook would
break if browsers didn't have consistent ordering so....

Test Plan: - `npm run test`

Reviewers: jlfwong, csilvers

Reviewed By: jlfwong, csilvers

Subscribers: alpert, csilvers, john

Differential Revision: https://phabricator.khanacademy.org/D24493