Update the aphrodite API to easily allow for server-side rendering.
Summary:
Update the aphrodite API by making the usage of rendering clearer. Namely:

 - `StyleSheet.renderBuffered` now handles buffered rendering on the client,
   as well as rehyrdation in an initial render.
 - `StyleSheetServer.renderStatic` now handles rendering on the server, and
   returns the appropriate data for rendering HTML and rehydrating.

I moved a bunch of the injection functions into a separate inject.js file to make
index.js cleaner and to allow for testing of some of the internal functions. I
added tests for the functions in inject.js, and a test to make sure renderStatic
returns what it's supposed to.

I added examples of server-side rendering in a second diff, D24515.

Test Plan: - `npm run test`

Reviewers: john, csilvers, jlfwong

Reviewed By: csilvers, jlfwong

Differential Revision: https://phabricator.khanacademy.org/D24509