Add an empty .npmignore to allow the lib/ dir to be published
Summary:
Since we want our package to be available for consumptino as an ES5 module, we
do a prepublish step and make the entry point of the package in lib instead of
src.

If you don't have a .npmignore file, npm will use your .gitignore, which we set
to ignore the lib directory, preventing it from being uploaded as part of the
publish.

Test Plan:
    npm publish

Then consume it in another repository.

Auditors: emily