Update inline-style-prefixer ^2.0.0 -> ^3.0.1
It seems that this version is a complete rewrite with some performance
improvements. In my benchmarks, this seems to improve the speed of
`css()` by about 10%.

Changelog:
* performance improvements (~10% faster)
* ordering prefixed properties correctly
* introducing new core prefixer that perform up to 4.5x faster
* added a whole new generator to create your custom prefixer
* added 4 new plugins to prefix special values
* new documentation using gitbook
* integrated flowtype

It seems that some of the prefixing is different, so I updated the tests
to match the new behavior. Since I had a hard time seeing the difference
in the test failures, I made them a little easier to see by adding some
formatting to the error messages.