Add bracket optimization to importantify
In our microbenchmark, adding this optimization performs a good amount
better. Since it is a simple optimization to add, it seems like we may
as well do it.

https://jsperf.com/slice-vs-bracket-string-check-with-imporant

I considered using bracket notation for the whole check, but I decided
that was too verbose. So instead, I decided to just check for one
character and follow up with a more thorough but concise check
afterward. This optimization favors the case where the style does not
already have `!important` on it, which is by and large likely to be the
general case.