Replace murmur hash with djb2 hash (#203)
* Replace murmur hash with djb2 hash

In profiling StyleSheet.create, I noticed that much of the time was
spent hashing. So, I found a faster hashing algorithm.

The implementation was taken from:

  https://github.com/darkskyapp/string-hash

According to this StackExchange post, this algorithm doesn't have as
good of randomness, but it has about the same percentage of collisions.
I don't think randomness matters for this application, so I think this
is okay.

  http://softwareengineering.stackexchange.com/a/145633

Using similar methodology to #202, this appears to make StylSheet.create
~15% faster (~220ms to ~185ms).

* Depend on string-hash for djb2 hashing algorithm

This is where I copied the code for this algorithm from, seems like we
might as well just bring in the dependency for it.