Re-enable prefixing for all browser versions
The v3 update of inline-style-prefixer changed the set of browsers that
it would generate prefixes for by default. To prevent this from being a
breaking change for Aphrodite users, we are generating our own prefixer
data in a way that supports all versions of all browsers. The guide for
this can be found at:

  https://github.com/rofrischmann/inline-style-prefixer/blob/master/docs/guides/CustomPrefixAll.md

It would be nice to eventually allow consumers of Aphrodite to customize
the browser support matrix, but in the meantime, this will do.