Handle null property values
Summary: The recent changes to `importantify` made it no longer worked when
passed in non-strings (such as `null` and `undefined`). This recovers the old
behaviour by manually converting the input to a string, and adds a test to make
sure this edge case is handled in the future.

Test Plan:
 - `npm run test`

@lencioni