Refactor out recursiveMerge
We only use this in one place, so I decided to rewrite this for the
specific use-case to improve performance.

While I was at it, I found and fixed a bug that causes nested objects to
be mutated.

I decided to change the signature of OrderedElements's forEach callback
to match that of Map's forEach.

In my benchmark this brings css() down from 1470ms to 1350ms.