Add coverage and .DS_Store to .gitignore/.eslintignore
The coverage directory can be generated to see more information about
test coverage. We don't want to check this in or enforce code quality
tools on it, so let's ignore it if it exists.

Similarly, MacOS has a tendency to drop .DS_Store files all over the
place, and we don't want to check these in.

While I was at it, I sorted these files alphabetically.