Inline mapObj into StyleSheet.create
Removing this abstraction that was only used in one place simplifies the
code a little and theoretically should improve performance by avoiding
some extra function calls and object creation. I say theoretically
because my rudimentary profiling is pretty inconclusive. It is possible
that my benchmark scenario is non-representative, the browser is
optimizing this away anyway. In either case, this seems like a small
improvement.