Replace object spread with a reducer
In my previous commit, I did this with a Set, but I was bummed about
this being a breaking change that causes people to have to make sure
they use a polyfill for older browsers, so I gave it some more thought.
After doing some profiling, I've found that this method may actually be
a little faster than Set, and it is not breaking. Seems like a good win.