Expose flushToStyleTag
v2 switched from webpack to rollup for its build. As part of this, the
individual files that were previously importable by intrepid spirits
gave access to some functions that were not explicitly exposed as part
of Aphrodite's public API.

At least one of these is cited in some issues as a workaround for some
edge cases like #76, and is used by projects like
react-with-styles-interface-aphrodite to allow folks to work around some
of these same types of issues.

To allow folks relying on this function to update to v2, I am adding
this as an export.