Update some packages to fix security warnings.
I had to fix a broken test that was asserting that a substring was
present in some element of a list.  I am guessing old versions of
mocha stringified the list and checked if the substring was in that
stringification, but now it gives a type error (as it should).

Test Plan:
I ran `npm test`.  This doesn't test coveralls -- I don't know how to
do that -- but even if it's wrong a github-bot should notice.  No
changes were made to library-specific deps so I feel pretty confident
this won't break things badly.